requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile '.\requirements.in'
#
aiofiles==24.1.0
    # via -r .\requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   starlette
asgiref==3.8.1
    # via -r .\requirements.in
async-timeout==5.0.1
    # via redis
certifi==2025.4.26
    # via
    #   httpcore
    #   httpx
cffi==1.17.1
    # via gevent
click==8.1.8
    # via uvicorn
colorama==0.4.6
    # via
    #   click
    #   loguru
dnspython==2.7.0
    # via pymongo
dramatiq[watch]==1.17.1
    # via -r .\requirements.in
exceptiongroup==1.2.2
    # via anyio
fastapi==0.115.12
    # via -r .\requirements.in
gevent==25.4.2
    # via watchdog-gevent
greenlet==3.2.1
    # via gevent
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via -r .\requirements.in
idna==3.10
    # via
    #   anyio
    #   httpx
loguru==0.7.3
    # via -r .\requirements.in
lxml==5.4.0
    # via python-docx
motor==3.7.0
    # via -r .\requirements.in
packaging==25.0
    # via pytesseract
pdf2image==1.17.0
    # via -r .\requirements.in
pillow==11.2.1
    # via
    #   -r .\requirements.in
    #   pdf2image
    #   pytesseract
prometheus-client==0.21.1
    # via dramatiq
pycparser==2.22
    # via cffi
pydantic==2.11.3
    # via fastapi
pydantic-core==2.33.1
    # via pydantic
pymongo==4.12.1
    # via
    #   -r .\requirements.in
    #   motor
pypdf2==3.0.1
    # via -r .\requirements.in
pytesseract==0.3.13
    # via -r .\requirements.in
python-docx==1.1.2
    # via -r .\requirements.in
python-dotenv==1.1.0
    # via -r .\requirements.in
python-multipart==0.0.20
    # via -r .\requirements.in
redis==4.6.0
    # via -r .\requirements.in
sniffio==1.3.1
    # via anyio
starlette==0.46.2
    # via fastapi
typing-extensions==4.13.2
    # via
    #   anyio
    #   asgiref
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   python-docx
    #   typing-inspection
    #   uvicorn
typing-inspection==0.4.0
    # via pydantic
uvicorn==0.34.2
    # via -r .\requirements.in
watchdog==6.0.0
    # via
    #   dramatiq
    #   watchdog-gevent
watchdog-gevent==0.2.1
    # via dramatiq
win32-setctime==1.2.0
    # via loguru
zope-event==5.0
    # via gevent
zope-interface==7.2
    # via gevent

# The following packages are considered to be unsafe in a requirements file:
# setuptools
